Move close wrapper in to general test helpers package
